Move api-examples -> mead/api_examples
this allows us to make entry points for setup
