+++ baseline/version.py
@@ -1,1 +1,1 @@
-__version__ = "2.3.15"
+__version__ = "2.3.16"
