+++ mead/api_examples/pretrain_discrim_tf.py
