+++ mead/api_examples/pretrain_tlm_tf.py
@@ -5,16 +5,17 @@ from argparse import ArgumentParser
 import math
 from typing import Tuple
 import baseline
-from eight_mile.utils import str2bool, write_json, read_json
+from eight_mile.utils import str2bool, write_json
 import baseline.tf.embeddings
 import baseline.embeddings
 from baseline.vectorizers import BPEVectorizer1D
-from eight_mile.utils import Average, get_num_gpus_multiworker, Offsets
-from eight_mile.tf.layers import create_distribute_strategy, TransformerEncoderStack, EmbeddingsStack, read_yaml_tf, read_json_tf
+from eight_mile.utils import Average, Timer, get_num_gpus_multiworker
+from eight_mile.tf.layers import create_distribute_strategy, read_yaml_tf
 from eight_mile.optz import *
 from eight_mile.tf.optz import *
-from baseline.tf.tfy import SET_TRAIN_FLAG
-from eight_mile.tf.serialize import save_tlm_output_npz
+from baseline.tf.lm import SET_TRAIN_FLAG, TransformerLanguageModel, TransformerMaskedLanguageModel
+from eight_mile.tf.serialize import save_tlm_npz
+from collections.abc import Mapping
 import tensorflow as tf
 import json
 logger = logging.getLogger(__file__)
@@ -30,40 +31,9 @@ Sample preprocessed data can be found here: https://www.dropbox.com/s/jir4layu6n
 """
 
 
-class TransformerTagger(tf.keras.Model):
-    def __init__(
-        self,
-        num_labels: int,
-        embeddings: tf.keras.layers.Layer,
-        num_heads=8, num_layers=8, d_model=512, d_ff=None, rpr_k=None, activation='gelu', scale=True, ffn_pdrop=0.0,
-        layer_norm_eps=1.0e-6, rpr_value_on=True, layer_drop=0.0, dropout=0.1,
-        name: str = None, **kwargs
-
-    ):
-        super().__init__(name=name)
-        self.embeddings = EmbeddingsStack(embeddings)
-        self.transformer = TransformerEncoderStack(
-            num_heads, d_model=d_model,
-            pdrop=dropout, scale=scale,
-            layers=num_layers, d_ff=d_ff, rpr_k=rpr_k,
-            activation=activation,
-            ffn_pdrop=ffn_pdrop,
-            layer_norm_eps=layer_norm_eps,
-            rpr_value_on=rpr_value_on,
-            layer_drop=layer_drop)
-        self.output_layer = tf.keras.layers.Dense(num_labels)
-
-    def call(self, inputs):
-        input_mask = tf.expand_dims(tf.expand_dims(tf.cast(inputs['x'] != Offsets.PAD, tf.int32), 1), 1)
-        embed = self.embeddings(inputs)
-        transformed = self.transformer((embed, input_mask))
-        output = self.output_layer(transformed)
-        return output
-
-
 def loss_function(model, features, labels):
-    logits = model(features)
-    loss_mask = tf.cast(labels != Offsets.PAD, tf.float32)
+    logits, _ = model(features, None)
+    loss_mask = tf.cast(labels != 0, tf.float32)
     losses = tf.nn.sparse_softmax_cross_entropy_with_logits(logits=logits, labels=labels)
     losses = losses * loss_mask
     losses = tf.reduce_sum(losses)
@@ -88,11 +58,16 @@ def _parse_tf_record(example_proto):
     return record['x'], record['y']
 
 
+def _parse_tf_record_causal(example_proto):
+    record = tf.io.parse_single_example(example_proto, feature_description)
+    return record['x'][:-1], record['x'][1:]
+
+
 def decode_json(example):
     return tf.py_function(_parse_json, [example], [tf.int32, tf.int32])
 
 
-def get_dataset(directory, file_type, num_parallel_reads=1, shuffle=True):
+def get_dataset(directory, file_type, num_parallel_reads=1, shuffle=True, causal=False):
     """Get a dataset as a tf.data.Dataset.  Input can be a bucket or a local file
 
 
@@ -100,6 +75,7 @@ def get_dataset(directory, file_type, num_parallel_reads=1, shuffle=True):
     :param file_type: Currently supports "json" files or "tfrecords"
     :param num_parallel_reads: The number of parallel reads
     :param shuffle: Defaults to True
+    :param causal: Use CLM instead of MLM (Defaults to ``False``)
     :return: a `tf.data.Dataset`
     """
     pattern = os.path.join(directory, f'*.{file_type}')
@@ -107,6 +83,8 @@ def get_dataset(directory, file_type, num_parallel_reads=1, shuffle=True):
     logger.debug(files)
 
     if file_type == 'json':
+        if causal:
+            raise Exception("Causal not currently supported with JSON input")
         ds = tf.data.TextLineDataset(files, num_parallel_reads=num_parallel_reads)
         if shuffle:
             ds = ds.shuffle(100)
@@ -121,7 +99,8 @@ def get_dataset(directory, file_type, num_parallel_reads=1, shuffle=True):
                            num_parallel_calls=tf.data.experimental.AUTOTUNE,
                            cycle_length=num_parallel_reads)
         ds = ds.shuffle(buffer_size=100)
-    ds = ds.map(_parse_tf_record)
+    parse_fn = _parse_tf_record_causal if causal else _parse_tf_record
+    ds = ds.map(parse_fn)
     return ds
 
 
@@ -133,14 +112,13 @@ def get_num_samples(sample_md):
 
 
 
-def train():
+def main():
     parser = ArgumentParser()
     parser.add_argument("--basedir", type=str)
     parser.add_argument("--train_dir", type=str, required=True, help='Training directory')
     parser.add_argument("--valid_dir", type=str, required=True, help='Validation directory')
     parser.add_argument("--train_md", type=str, help="Training metadata YAML, defaults to `{train_dir}/md.yml`")
     parser.add_argument("--valid_md", type=str, help="Validation metadata YAML, defaults to `{valid_dir}/md.yml`")
-    parser.add_argument("--label_file", type=str, help="JSON file mapping labels to integers", default="labels.json")
     parser.add_argument("--dataset_key", default="tlm",
                         help="dataset key for basedir")
     parser.add_argument("--embed_type", type=str, default='default',
@@ -148,6 +126,7 @@ def train():
                         help="register label of the embeddings")
     parser.add_argument("--d_model", type=int, default=512, help="Model dimension (and embedding dsz)")
     parser.add_argument("--d_ff", type=int, default=2048, help="FFN dimension")
+    parser.add_argument("--d_k", type=int, default=None, help="Dimension per head.  Use if num_heads=1 to reduce dims")
     parser.add_argument("--num_heads", type=int, default=8, help="Number of heads")
     parser.add_argument("--num_layers", type=int, default=8, help="Number of layers")
     parser.add_argument("--num_train_workers", type=int, default=4, help="Number train workers")
@@ -168,6 +147,7 @@ def train():
     parser.add_argument("--epochs", type=int, default=32, help="Num training epochs")
     parser.add_argument("--restart", type=str2bool, help="Option allows you to restart from a previous checkpoint")
     parser.add_argument("--warmup_steps", type=int, default=10000, help="Num warmup steps")
+    parser.add_argument("--causal", type=str2bool, default=False, help="Use CLM (causal) instead of MLM")
     parser.add_argument("--saves_per_epoch", type=int, default=10, help="The number of checkpoints to save per epoch")
     parser.add_argument('--rpr_k',
                         help='Relative attention positional sizes pass 0 if you dont want relative attention',
@@ -192,7 +172,7 @@ def train():
     logger.info(f"Writing results to {args.basedir}")
 
     if args.tb:
-        logdir = f"logs/scalars/{os.getpid()}"
+        logdir = f"{args.basedir}/scalars/{os.getpid()}"
         file_writer = tf.summary.create_file_writer(logdir + "/metrics")
         file_writer.set_as_default()
         logger.info(f"Set up tensorboard logdir {logdir}")
@@ -206,18 +186,32 @@ def train():
                                                        preserve_vocab_indices=True,
                                                        embed_type=args.embed_type)
     vocabs = preproc_data['vocab']
+    vocab_size = max(vocabs.values())
+
 
     train_md = args.train_md if args.train_md else os.path.join(args.train_dir, 'md.yml')
     num_train_samples = get_num_samples(train_md)
     valid_md = args.valid_md if args.valid_md else os.path.join(args.valid_dir, 'md.yml')
     num_valid_samples = get_num_samples(valid_md)
-    labels = read_json_tf(args.label_file)
-    num_labels = len(labels)
+
+    is_curriculum = True if isinstance(num_train_samples, Mapping) else False
 
     def dataset_train_fn(input_context):
         global_batchsz = args.batch_size
         base_batchsz = input_context.get_per_replica_batch_size(global_batchsz)
-        ds = get_dataset(args.train_dir, args.file_type, args.num_train_workers).batch(base_batchsz)
+        ds = None
+        if is_curriculum:
+            for sub in num_train_samples.keys():
+                train_curr_dir = os.path.join(args.train_dir, str(sub))
+                batchsz_scale_factor = args.nctx // sub
+                this_batchsz = base_batchsz * batchsz_scale_factor
+                curr_ds = get_dataset(train_curr_dir, args.file_type, args.num_train_workers, causal=args.causal).batch(this_batchsz, drop_remainder=True)
+                if ds is None:
+                    ds = curr_ds
+                else:
+                    ds = ds.concatenate(curr_ds)
+        else:
+            ds = get_dataset(args.train_dir, args.file_type, args.num_train_workers, causal=args.causal).batch(base_batchsz)
         return ds.shard(
             input_context.num_input_pipelines, input_context.input_pipeline_id
         )
@@ -226,7 +220,20 @@ def train():
     def dataset_test_fn(input_context):
         global_batchsz = args.batch_size
         base_batchsz = input_context.get_per_replica_batch_size(global_batchsz)
-        ds = get_dataset(args.valid_dir, args.file_type, args.num_train_workers, shuffle=False).batch(base_batchsz)
+        ds = None
+        if is_curriculum:
+            for sub in num_valid_samples.keys():
+                valid_curr_dir = os.path.join(args.valid_dir, str(sub))
+                batchsz_scale_factor = args.nctx // sub
+                this_batchsz = base_batchsz * batchsz_scale_factor
+                curr_ds = get_dataset(valid_curr_dir, args.file_type, args.num_train_workers, causal=args.causal).batch(
+                    this_batchsz, drop_remainder=True)
+                if ds is None:
+                    ds = curr_ds
+                else:
+                    ds = ds.concatenate(curr_ds)
+        else:
+            ds = get_dataset(args.valid_dir, args.file_type, args.num_train_workers, shuffle=False, causal=args.causal).batch(base_batchsz)
 
         return ds.shard(
             input_context.num_input_pipelines, input_context.input_pipeline_id
@@ -242,16 +249,42 @@ def train():
     logger.info("Loaded datasets")
     logger.info("Using embedding type [%s]", args.embed_type)
     if len(args.rpr_k) == 0 or args.rpr_k[0] < 1:
-        args.rpr_k = None
+        rpr_k = None
     elif len(args.rpr_k) == 1:
-        args.rpr_k = args.rpr_k[0]
+        rpr_k = args.rpr_k[0]
+    else:
+        rpr_k = args.rpr_k
 
-    model = TransformerTagger(num_labels, embeddings, **vars(args))
+    TLM = TransformerLanguageModel if args.causal else TransformerMaskedLanguageModel
+    model = TLM.create(embeddings,
+                       hsz=args.d_model,
+                       d_ff=args.d_ff,
+                       tie_weights=True,
+                       dropout=args.dropout,
+                       gpu=False,
+                       num_heads=args.num_heads,
+                       layers=args.num_layers,
+                       rpr_k=rpr_k,
+                       d_k=args.d_k,
+                       ffn_pdrop=args.ffn_pdrop,
+                       windowed_ra=args.windowed_ra,
+                       rpr_value_on=args.rpr_value_on,
+                       layer_drop=args.layer_drop,
+                       src_keys=['x'], tgt_key='x')
 
     logger.info("Loaded model and loss")
 
-    steps_per_epoch = num_train_samples // args.batch_size
-    steps_per_valid_epoch = num_valid_samples // args.batch_size
+    if is_curriculum:
+        steps_per_epoch = 0
+        steps_per_valid_epoch = 0
+        for k, v in num_train_samples.items():
+            steps_per_epoch += int(num_train_samples[k] // (args.batch_size * (args.nctx / k)))
+        for k, v in num_valid_samples.items():
+            steps_per_valid_epoch += int(num_valid_samples[k] // (args.batch_size * (args.nctx / k)))
+
+    else:
+        steps_per_epoch = num_train_samples // args.batch_size
+        steps_per_valid_epoch = num_valid_samples // args.batch_size
     update_on = steps_per_epoch // args.saves_per_epoch
     report_on = max(10, update_on) // 10
     logger.info(f"Steps per epoch: {steps_per_epoch}. Saving checkpoint every {update_on} steps.")
@@ -309,11 +342,11 @@ def train():
     with strategy.scope():
 
         for epoch in range(start_epoch, args.epochs):
+            timer.start()
             SET_TRAIN_FLAG(True)
             logger.info('Starting epoch %d', epoch + 1)
             avg_loss = Average('average_train_loss')
             metrics = {}
-            timer.start()
             train_iter = iter(train_loader)
             for i in range(steps_per_epoch):
 
@@ -328,7 +361,7 @@ def train():
                     logger.warning("Convert only flag specified.  Stopping after one step")
                     steps = optimizer.global_step.numpy()
                     npz_checkpoint = os.path.join(args.basedir, f'checkpoint-step-{steps}.npz')
-                    save_tlm_output_npz(model, npz_checkpoint)
+                    save_tlm_npz(model, npz_checkpoint)
                     return
 
                 steps = optimizer.global_step.numpy()
@@ -342,15 +375,14 @@ def train():
                     if args.npz:
 
                         npz_checkpoint = os.path.join(args.basedir, f'checkpoint-step-{steps}.npz')
-                        save_tlm_output_npz(model, npz_checkpoint)
+                        save_tlm_npz(model, npz_checkpoint)
 
             # How much time elapsed in minutes
-            elapsed = timer.elapsed(True)
             train_token_loss = avg_loss.avg
             # This is the average training token-level loss across all machines
             # This is the token-level training perplexity
             train_token_ppl = math.exp(train_token_loss)
-            metrics['train_elapsed_min'] = elapsed
+            metrics['train_elapsed_min'] = timer.elapsed(True)
             metrics['average_train_loss'] = train_token_loss
             metrics['train_ppl'] = train_token_ppl
             metrics['lr'] = float(lr_sched(tf.cast(optimizer.global_step, tf.float32)).numpy().item())
@@ -370,15 +402,12 @@ def train():
 
             valid_token_loss = avg_valid_loss.avg
             valid_token_ppl = math.exp(valid_token_loss)
-
-            elapsed = timer.elapsed(True)
-
-            metrics['valid_elapsed_min'] = elapsed
+            metrics['valid_elapsed_min'] = timer.elapsed(True)
             metrics['average_valid_loss'] = valid_token_loss
             metrics['average_valid_word_ppl'] = valid_token_ppl
             logger.info(json.dumps(metrics, indent=4))
 
 
 if __name__ == "__main__":
-    train()
+    main()
